POC for csv export polling
This is a POC MR to evaluate polling for CSV export.
When a merge commit csv is requested, a background job
is created that computes, generates CSV and maintains the
status of the export.
The UI meanwhile polls to see if an export is ready,
and shows the download button accordingly.