Limit trace length when called from Pipeline.jobs
This changes the behaviour of `CiJob.trace.raw` and `CiJob.trace.html`
so that these fields behave differently in different contexts.

When called as part of a collection (eg. as `Pipeline.jobs.trace.raw`)
then:

- the maximum trace length is 25 lines
- the client is forbidden from requesting the full trace

When called on a single job (eg. as `Pipeline.job(id).trace.raw`):

- There is no maximum tail length
- the client is allowed to request the full trace

The aim of this is to limit the potential abuse of this feature, but
limited the size of traces that are loaded into memory in the course of
a query.

At the moment there is no way to request a single job, which is
implemented in https://gitlab.com/gitlab-org/gitlab/-/merge_requests/44703