Place hard limits on the number of build-traces per-query
This adds a limit on the number of instances of `BuildTrace` that any
one query may return. This aims to prevent resource exhaustion on the
server by loading many traces in a large pipeline.

The limit is evaluated before any trace is loaded from storage, so if
the limit is breached, we will not need to do any unnecessary IO.

In tandem with the variable tail-length limit (set to 25 lines at most
when fetched from a collection), this limits the scope for abuse of this
feature.