POC: Avoid parsing blobs to highlight diffs
Currently, we load and highlight old and new versions
of the blob of a diff. Then diff lines are mapped to
the highlighted content of the blob. It can be quite
expensive when a small part of a huge file is changed.

Let's instead extract blob content from diff lines and
highlight them.

For example, when we have

```
 unchanged line
 +added line
 -removed line
 ...
 another unchanged line
 -another added line
```

We parse it into old content chunks:
[[removed line], []]

And new content chunks:
[
  [unchanged line, added line],
  [another unchanged line, another added line]
]

Then we highlight each chunk separately and map the
diff lines to the content of the chunks.