POC: For GraphQL based list
This MR is a proof pf concept of what the runners list would look like
if we used a format similar to the one used for issues (and merge
requests).

It uses the familiar patterns so administrators can find easily what
they are looking and filter the runner they need.