Cache objects being used in note policy checks in request
When we request for discussions, we also filter the notes we show
if they're readable by user via policy checks.

When there are a number of notes (e.g. 100+) in discussions, it
means that the policy check will run the same number of times.

The policies involved with it seems to instantiate the same objects
for each note (e.g. noteable, project).

This reduces those objects by caching it so the same object will
be reused. The cache is scoped by request since it only needs to
happen on the same request.

Changelog: performance
EE: true