Fix cache key for QA gems for Debian buster
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/80362 we
bumped the CI images from Debian buster to bullseye. We bumped the
cache key for everything but the QA gems, which appear to be cached in
the `setup-test-env` job.

Since Debian bullseye uses libc 2.29 but buster uses 2.28, this led to
errors in the `14-7-stable-ee` branch running C extensions compiled
against the former.

To avoid this problem in the short-term, bump the cache key in
`14-7-stable-ee` while we also update the cache key in `master`
(https://gitlab.com/gitlab-org/gitlab/-/merge_requests/80656).