Use OAuth authentication for Jira connect app
By changing the sign in method from cookies to OAuth,
we will enable self-hosted users to use the GitLab for Jira
app without any specific workarounds by using GitLab.com
as a proxy. This will be added in the next iteration of
this feature and is currently behind a feature flag.

Changelog: changed